Изменил тип поля

+++ eko-server/src/main/java/org/ekolab/server/model/content/lab3/Lab3Data.java
@@ -75,7 +75,7 @@ public class Lab3Data extends LabData<Lab3Variant> {
     /**
      * Расход топлива на 1 блок
      */
-    private Double fuelConsumer;
+    private Integer fuelConsumer;
 
     /**
      * Потери тепла от механической неполноты сгорания топлива
@@ -442,11 +442,11 @@ public class Lab3Data extends LabData<Lab3Variant> {
         this.lowHeatValue = lowHeatValue;
     }
 
-    public Double getFuelConsumer() {
+    public Integer getFuelConsumer() {
         return fuelConsumer;
     }
 
-    public void setFuelConsumer(Double fuelConsumer) {
+    public void setFuelConsumer(Integer fuelConsumer) {
         this.fuelConsumer = fuelConsumer;
     }
